requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==20.7
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
alembic==1.13.0
greenlet==3.2.4
pytest==7.4.3
pytest-asyncio==0.21.1
